refactor: extract internal/uitheme; onboarding no longer imports tui
- New internal/uitheme package owns WaveTheme, SelectionStyle,
  WaveLogo. Lightweight, free of bubbletea-specific deps so
  non-TUI callers (notably internal/onboarding) can consume it
  without pulling the full TUI runtime.
- internal/onboarding/{steps,ontology_step}.go imports
  internal/uitheme instead of internal/tui.
- internal/tui retains thin alias shim for existing consumers
  (theme.go); migration to internal/uitheme is opt-in and
  non-breaking.

Closes #1497 (residual). Smoke-contracts pipeline fix deferred
to follow-up.
